+++ src/main/java/edu/gcsc/celltreeedit/ClusterAnalysis/RelPartitioningErrorTable.java
@@ -17,7 +17,7 @@ public class RelPartitioningErrorTable extends JPanel implements Serializable {
 
     JTable table;
 
-    public RelPartitioningErrorTable(float[][] relPartitioningErrors, List<UniqueMetadataContainer.UniqueMetadata> uniqueMetadata) {
+    public RelPartitioningErrorTable(double[][] relPartitioningErrors, List<UniqueMetadataContainer.UniqueMetadata> uniqueMetadata) {
 
         // decimal format and rounding
         DecimalFormat df = new DecimalFormat("#.##");
